+++ src/main/java/backend/stub/DummyRepoState.java
@@ -36,7 +36,8 @@ public class DummyRepoState {
     // getEvents and getComments).
     private HashMap<Integer, IssueMetadata> issueMetadata = new HashMap<>();
     // We keep track of issues that user has not gotten metadata from.
-    private HashSet<Integer> updatedMetadata = new HashSet<>();
+    private HashSet<Integer> updatedEvents = new HashSet<>();
+    private HashSet<Integer> updatedComments = new HashSet<>();
 
     public DummyRepoState(String repoId) {
         this.dummyRepoId = repoId;
@@ -103,12 +104,12 @@ public class DummyRepoState {
         Comment[] ownComments = { ownComment };
         issues.get(9).setCommentCount(1);
         issues.get(9).setUpdatedAt(LocalDateTime.now());
-        issueMetadata.put(9, new IssueMetadata(
+        issueMetadata.put(9, IssueMetadata.intermediate(
                 new ArrayList<>(),
                 new ArrayList<>(Arrays.asList(ownComments)),
-                ""
+                "", ""
         ));
-        updatedMetadata.add(9);
+        updatedEvents.add(9);
 
         // Then put down three comments for issue 10
         Comment dummyComment1 = new Comment();
@@ -123,12 +124,12 @@ public class DummyRepoState {
         Comment[] dummyComments = { dummyComment1, dummyComment2, dummyComment3 };
         issues.get(10).setCommentCount(3);
         issues.get(10).setUpdatedAt(LocalDateTime.now());
-        issueMetadata.put(10, new IssueMetadata(
+        issueMetadata.put(10, IssueMetadata.intermediate(
                 new ArrayList<>(),
                 new ArrayList<>(Arrays.asList(dummyComments)),
-                ""
+                "", ""
         ));
-        updatedMetadata.add(10);
+        updatedEvents.add(10);
     }
 
     protected ImmutableTriple<List<TurboIssue>, String, Date>
@@ -138,7 +139,7 @@ public class DummyRepoState {
         if (!updatedIssues.isEmpty() || eTag == null) currETag = UUID.randomUUID().toString();
 
         ImmutableTriple<List<TurboIssue>, String, Date> toReturn = new ImmutableTriple<>(
-            new ArrayList<>(updatedIssues.values()), currETag, lastCheckTime);
+            deepCopyIssues(updatedIssues), currETag, lastCheckTime);
 
         updatedIssues = new TreeMap<>();
         return toReturn;
@@ -149,7 +150,7 @@ public class DummyRepoState {
         if (!updatedLabels.isEmpty() || eTag == null) currETag = UUID.randomUUID().toString();
 
         ImmutablePair<List<TurboLabel>, String> toReturn
-            = new ImmutablePair<>(new ArrayList<>(updatedLabels.values()), currETag);
+            = new ImmutablePair<>(deepCopyLabels(updatedLabels), currETag);
 
         updatedLabels = new TreeMap<>();
         return toReturn;
@@ -160,7 +161,7 @@ public class DummyRepoState {
         if (!updatedMilestones.isEmpty() || eTag == null) currETag = UUID.randomUUID().toString();
 
         ImmutablePair<List<TurboMilestone>, String> toReturn
-            = new ImmutablePair<>(new ArrayList<>(updatedMilestones.values()), currETag);
+            = new ImmutablePair<>(deepCopyMilestones(updatedMilestones), currETag);
 
         updatedMilestones = new TreeMap<>();
         return toReturn;
@@ -171,26 +172,26 @@ public class DummyRepoState {
         if (!updatedUsers.isEmpty() || eTag == null) currETag = UUID.randomUUID().toString();
 
         ImmutablePair<List<TurboUser>, String> toReturn
-            = new ImmutablePair<>(new ArrayList<>(updatedUsers.values()), currETag);
+            = new ImmutablePair<>(deepCopyUsers(updatedUsers), currETag);
 
         updatedUsers = new TreeMap<>();
         return toReturn;
     }
 
     protected List<TurboIssue> getIssues() {
-        return new ArrayList<>(issues.values());
+        return deepCopyIssues(issues);
     }
 
     protected List<TurboLabel> getLabels() {
-        return new ArrayList<>(labels.values());
+        return deepCopyLabels(labels);
     }
 
     protected List<TurboMilestone> getMilestones() {
-        return new ArrayList<>(milestones.values());
+        return deepCopyMilestones(milestones);
     }
 
     protected List<TurboUser> getCollaborators() {
-        return new ArrayList<>(users.values());
+        return deepCopyUsers(users);
     }
 
     private TurboIssue makeDummyIssue() {
@@ -224,15 +225,15 @@ public class DummyRepoState {
     }
 
     protected ImmutablePair<List<TurboIssueEvent>, String> getEvents(int issueId, String currentETag) {
-        if (updatedMetadata.contains(issueId)) {
+        if (updatedEvents.contains(issueId)) {
             IssueMetadata metadataOfIssue = issueMetadata.get(issueId);
 
-            // If updatedMetadata contains issue, its metaadata must also exist in the metadataOfIssue hashmap
+            // If updatedEvents contains issue, its metadata must also exist in the metadataOfIssue hashmap
             assert metadataOfIssue != null;
 
-            // Remove issue from updatedMetadata so that next time metadata is retrievd, the same ETag
+            // Remove issue from updatedEvents so that next time metadata is retrieved, the same ETag
             // will not be sent again unless more updates will have been introduced.
-            updatedMetadata.remove(issueId);
+            updatedEvents.remove(issueId);
 
             // Finally, return events as a proper array.
             return new ImmutablePair<>(metadataOfIssue.getEvents(), UUID.randomUUID().toString());
@@ -241,12 +242,11 @@ public class DummyRepoState {
     }
 
     protected List<Comment> getComments(int issueId) {
+        // TODO after implementing support for comments ETags on GitHubClientExtended, change logic to getEvents'.
         IssueMetadata metadataOfIssue = issueMetadata.get(issueId);
-
         if (metadataOfIssue != null) {
-            return metadataOfIssue.getComments();
+            return new ArrayList<>(metadataOfIssue.getComments());
         }
-        // If not in the metadata hashmap yet, there are currently no comments for given issue.
         return new ArrayList<>();
     }
 
@@ -290,7 +290,7 @@ public class DummyRepoState {
         toRename.setUpdatedAt(LocalDateTime.now());
 
         // Replace originals with copies, and queue them up to be retrieved
-        markUpdatedIssue(toRename, new IssueMetadata(eventsOfIssue, metadataOfIssue.getComments(), ""));
+        markUpdatedEvents(toRename, IssueMetadata.intermediate(eventsOfIssue, metadataOfIssue.getComments(), "", ""));
 
         return toRename;
     }
@@ -359,7 +359,7 @@ public class DummyRepoState {
         toSet.setUpdatedAt(LocalDateTime.now());
 
         // Replace originals with copies, and queue them up to be retrieved
-        markUpdatedIssue(toSet, new IssueMetadata(eventsOfIssue, metadataOfIssue.getComments(), ""));
+        markUpdatedEvents(toSet, IssueMetadata.intermediate(eventsOfIssue, metadataOfIssue.getComments(), "", ""));
 
         return labels.stream().map(new Label()::setName).collect(Collectors.toList());
     }
@@ -381,7 +381,8 @@ public class DummyRepoState {
         toComment.setCommentCount(toComment.getCommentCount() + 1);
 
         // Replace originals with copies, and queue them up to be retrieved
-        markUpdatedIssue(toComment, new IssueMetadata(metadataOfIssue.getEvents(), commentsOfIssue, ""));
+        markUpdatedComments(toComment,
+            IssueMetadata.intermediate(metadataOfIssue.getEvents(), commentsOfIssue, "", ""));
 
         return toComment;
     }
@@ -404,23 +405,94 @@ public class DummyRepoState {
                 metadataOfIssue.getComments() :
                 new ArrayList<>();
 
-        return new ImmutablePair<>(toMutate, new IssueMetadata(eventsOfIssue, commentsOfIssue, ""));
+        return new ImmutablePair<>(toMutate, IssueMetadata.intermediate(eventsOfIssue, commentsOfIssue, "", ""));
     }
 
     /**
-     * Auxiliary method that replaces original issue/metadata with the mutated copies after updating the issue.
-     * Simulates the event of an user action causing the ETag of the issue to change.
+     * Auxiliary method that replaces original issue & metadata with the mutated copies after updating the issue.
+     * Simulates the event of an user action causing the ETag of the issue events to change.
      *
      * @param toMark The mutated copy of the issue, to replace the original issue
      * @param toInsert The mutated metadata of the issue, to replace the original issue metadata
      */
-    private void markUpdatedIssue(TurboIssue toMark, IssueMetadata toInsert) {
+    private void markUpdatedEvents(TurboIssue toMark, IssueMetadata toInsert) {
         int issueId = toMark.getId();
 
         issues.put(issueId, toMark);
         updatedIssues.put(issueId, toMark);
         issueMetadata.put(issueId, toInsert);
-        updatedMetadata.add(issueId);
+        updatedEvents.add(issueId);
     }
 
+    /**
+     * Auxiliary method that replaces original issue & metadata with the mutated copies after updating the issue.
+     * Simulates the event of an user action causing the ETag of the issue comments to change.
+     *
+     * @param toMark The mutated copy of the issue, to replace the original issue
+     * @param toInsert The mutated metadata of the issue, to replace the original issue metadata
+     */
+    private void markUpdatedComments(TurboIssue toMark, IssueMetadata toInsert) {
+        int issueId = toMark.getId();
+
+        issues.put(issueId, toMark);
+        updatedIssues.put(issueId, toMark);
+        issueMetadata.put(issueId, toInsert);
+        updatedComments.add(issueId);
+    }
+
+    /**
+     * Copies the TreeMap of issues by creating a List containing copies of state-stored issues. Prevents
+     * external mutation of the issue objects from propagating to the repo state.
+     *
+     * @param issuesToCopy A TreeMap containing state-stored issues to copy from.
+     * @return A list containing copies of given issues.
+     */
+    private List<TurboIssue> deepCopyIssues(TreeMap<Integer, TurboIssue> issuesToCopy) {
+        ArrayList<TurboIssue> copiedIssues = new ArrayList<>();
+        issuesToCopy.values().forEach(issue -> copiedIssues.add(new TurboIssue(issue)));
+
+        return copiedIssues;
+    }
+
+    /**
+     * Copies the TreeMap of labels by creating a List containing copies of state-stored labels. Prevents
+     * external mutation of the label objects from propagating to the repo state.
+     *
+     * @param labelsToCopy A TreeMap containing state-stored labels to copy from.
+     * @return A list containing copies of given labels.
+     */
+    private List<TurboLabel> deepCopyLabels(TreeMap<String, TurboLabel> labelsToCopy) {
+        ArrayList<TurboLabel> copiedLabels = new ArrayList<>();
+        labelsToCopy.values().forEach(label -> copiedLabels.add(new TurboLabel(label)));
+
+        return copiedLabels;
+    }
+
+    /**
+     * Copies the TreeMap of milestones by creating a List containing copies of state-stored milestones. Prevents
+     * external mutation of the milestone objects from propagating to the repo state.
+     *
+     * @param milestonesToCopy A TreeMap containing state-stored milestones to copy from.
+     * @return A list containing copies of given milestones.
+     */
+    private List<TurboMilestone> deepCopyMilestones(TreeMap<Integer, TurboMilestone> milestonesToCopy) {
+        ArrayList<TurboMilestone> copiedMilestones = new ArrayList<>();
+        milestonesToCopy.values().forEach(milestone -> copiedMilestones.add(new TurboMilestone(milestone)));
+
+        return copiedMilestones;
+    }
+
+    /**
+     * Copies the TreeMap of users by creating a List containing copies of state-stored users. Prevents
+     * external mutation of the user objects from propagating to the repo state.
+     *
+     * @param usersToCopy A TreeMap containing state-stored users to copy from.
+     * @return A list containing copies of given users.
+     */
+    private List<TurboUser> deepCopyUsers(TreeMap<String, TurboUser> usersToCopy) {
+        ArrayList<TurboUser> copiedUsers = new ArrayList<>();
+        usersToCopy.values().forEach(user -> copiedUsers.add(new TurboUser(user)));
+
+        return copiedUsers;
+    }
 }