+++ src/main/java/github/update/MilestoneUpdateService.java
@@ -1,7 +1,7 @@
 package github.update;
 
 import com.google.gson.reflect.TypeToken;
-import github.GitHubClientExtended;
+import github.GitHubClientEx;
 import org.eclipse.egit.github.core.IRepositoryIdProvider;
 import org.eclipse.egit.github.core.Milestone;
 import org.eclipse.egit.github.core.client.PagedRequest;
@@ -12,10 +12,10 @@ import java.util.Map;
 
 import static org.eclipse.egit.github.core.client.IGitHubConstants.SEGMENT_MILESTONES;
 
-public class MilestoneUpdateService extends UpdateService<Milestone>{
+public class MilestoneUpdateService extends UpdateService<Milestone> {
 
 
-    public MilestoneUpdateService(GitHubClientExtended client, String milestonesETag){
+    public MilestoneUpdateService(GitHubClientEx client, String milestonesETag){
         super(client, SEGMENT_MILESTONES, milestonesETag);
     }