+++ src/main/java/ui/MenuControl.java
@@ -104,7 +104,7 @@ public class MenuControl extends MenuBar {
         createLeft.setOnAction(e -> {
             logger.info("Menu: Panels > Create (Left)");
             panels.createNewPanelAtStart();
-            setHvalue(panelsScrollPane.getHmin());
+            panels.scrollToCurrentlySelectedPanel();
         });
         createLeft.setAccelerator(CREATE_LEFT_PANEL);
 
@@ -122,7 +122,7 @@ public class MenuControl extends MenuBar {
                             ScrollBar scrollBar = (ScrollBar) child;
                             if (scrollBar.getOrientation() == Orientation.HORIZONTAL
                                     && scrollBar.visibleProperty().get()) {
-                                setHvalue(panelsScrollPane.getHmax());
+                                panels.scrollToCurrentlySelectedPanel();
                                 break;
                             }
                         }
@@ -366,13 +366,5 @@ public class MenuControl extends MenuBar {
 
     private void onRepoRemove(String repoId) {
         ui.logic.removeRepository(repoId).thenRun(() -> ui.triggerEvent(new OpenReposChangedEvent()));
-    }
-
-    public void scrollTo(int panelIndex, int numOfPanels){
-        setHvalue(panelIndex * (panelsScrollPane.getHmax()) / (numOfPanels - 1));
-    }
-
-    private void setHvalue(double val) {
-        panelsScrollPane.setHvalue(val);
     }
 }