+++ src/main/java/ui/components/NavigableListView.java
@@ -158,11 +158,11 @@ public abstract class NavigableListView<T> extends ScrollableListView<T> {
             }
             if (FIRST_ISSUE.match(e)) {
                 e.consume();
-                selectFirstItem();
+                selectNthItem(1);
             }
             if (LAST_ISSUE.match(e)) {
                 e.consume();
-                selectLastItem();
+                selectNthItem(getItems().size());
             }
         });
     }
@@ -188,21 +188,17 @@ public abstract class NavigableListView<T> extends ScrollableListView<T> {
         onItemSelected = callback;
     }
 
-    public void selectFirstItem() {
-        requestFocus();
-        if (getItems().size() == 0) return;
-        getSelectionModel().clearAndSelect(0);
-        scrollAndShow(0);
-        selectedIndex = Optional.of(0);
-        onItemSelected.accept(selectedIndex.get());
-    }
+    /**
+     * Selects the nth item in the navigable list.
+     * @param n - Nth item, indexing is 1-based (1st item would correspond to n = 1)
+     */
 
-    public void selectLastItem() {
+    public void selectNthItem(int n) {
         requestFocus();
-        if (getItems().size() == 0) return;
-        getSelectionModel().clearAndSelect(getItems().size() - 1);
-        scrollAndShow(getItems().size() - 1);
-        selectedIndex = Optional.of(getItems().size() - 1);
+        assert n <= getItems().size() && n > 0;
+        getSelectionModel().clearAndSelect(n - 1);
+        scrollAndShow(n - 1);
+        selectedIndex = Optional.of(n - 1);
         onItemSelected.accept(selectedIndex.get());
     }